PDb_Aux_Plugin class added $settings_API_status property to indicate whether the subclass is using the Settings API or not settings pages are not added for aux plugins that don't use the API, but it's not automatic, the property has to be set to false in the child class
PDb_FormElement class
the "pdb-before_display_field" filter was changed, now it only passes in the blank string and the whole field object...this was a necessary change that will break anyone's use of this filter, but it was not passing in enough info to properly implement a custom field before.


PDb_List_Admin class
the default case for value displays now uses the PDb_FormElement::get_field_value_display method so it can be filtered

xnau_FormElement class
null select options are no longer automatically added to fields that have no set value. This gave the designer no flexibility in setting up fields, because you could not prevent the inclusion of the null select option. Now, if there is no set value for a field and the null_select has been set to false, the field will have no value in the form.


git-svn-id: http://plugins.svn.wordpress.org/participants-database/trunk@965863 b8457f37-d9ea-0310-8a92-e5e31aec5664